Issue #862 Extensibility for action invoker
