+++ src/Microsoft.AspNet.Mvc.Core/IActionInvoker.cs
@@ -7,6 +7,6 @@ namespace Microsoft.AspNet.Mvc
 {
     public interface IActionInvoker
     {
-        Task InvokeActionAsync();
+        Task InvokeAsync();
     }
 }
